Abstract database and switch to mysql
While the sqlite3 database provides a nice clean way to have a
local database, there's no reason to not continue using the
mariadb instance running on the home network.

First we abstract out all the database calls to a separate
module/class. Then we switch the database command to work with
pymysql instead of sqlite3.

With these changes, we can now use the "routes" database that
already exists.

If in the future, we want to go back to using sqlite3, we just
need to write a new database module.